refactor: rive, takePhoto 커스텀훅으로 파일 분리
